@@ -18,21 +18,28 @@
     "@radix-ui/react-select": "^2.1.2",
     "@radix-ui/react-slot": "^1.1.0",
     "@react-google-maps/api": "^2.20.3",
+    "bcrypt": "^5.1.0",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
+    "esm": "^3.2.25",
+    "framer-motion": "^11.11.17",
     "lucide-react": "^0.454.0",
     "next": "^15.0.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-icons": "^5.3.0",
     "shadcn": "^1.0.0",
+    "sqlite": "^5.0.1",
+    "sqlite3": "^5.1.6",
+    "styled-jsx": "^5.1.6",
     "tailwind-merge": "^2.5.4",
     "tailwindcss-animate": "^1.0.7"
   },
   "devDependencies": {
-    "@types/node": "22.8.2",
+    "@types/node": "^22.9.0",
     "@types/react": "18.3.12",
-    "eslint": "^8",
-    "eslint-config-next": "15.0.1",
+    "eslint": "^9.14.0",
+    "eslint-config-next": "^15.0.2",
     "postcss": "^8",
     "tailwindcss": "^3.4.1",
     "typescript": "5.6.3"
