@@ -1,112 +1,189 @@
-'use client'
+"use client";
 
-import React from 'react'
-import { Button } from "@/components/ui/button"
-import { Card, CardContent } from "@/components/ui/card"
-import Header from "@/components/ui/pageHeader"
-import BottomNavBar from "@/components/ui/navBar"
-import { useRouter } from 'next/navigation'
-import { CreditCard } from 'lucide-react'
+import React from "react";
+import { Button } from "@/components/ui/button";
+import { Card, CardContent } from "@/components/ui/card";
+import Header from "@/components/ui/pageHeader";
+import BottomNavBar from "@/components/ui/navBar";
+import { useRouter } from "next/navigation";
+import { CreditCard, Check, Info } from "lucide-react";
 
 interface MealPlanOption {
-  id: string
-  semesterPrice: number
-  yearPrice: number
-  mealsPerWeek: number
+  id: string;
+  semesterPrice: number;
+  yearPrice: number;
+  mealsPerWeek: number;
+  description: string;
 }
 
 export default function OfferedMealPlansPage() {
-  const router = useRouter()
-  const [selectedTypes, setSelectedTypes] = React.useState<Record<string, 'Year' | 'Semester'>>({})
+  const router = useRouter();
+  const [selectedTypes, setSelectedTypes] = React.useState<
+    Record<string, "Year" | "Semester">
+  >({});
 
   const mealPlans: MealPlanOption[] = [
-    { id: '1', semesterPrice: 1299.00, yearPrice: 2499.00, mealsPerWeek: 12 },
-    { id: '2', semesterPrice: 1599.00, yearPrice: 3099.00, mealsPerWeek: 24 },
-    { id: '3', semesterPrice: 1399.00, yearPrice: 2699.00, mealsPerWeek: 12 },
-    { id: '4', semesterPrice: 1499.00, yearPrice: 2899.00, mealsPerWeek: 12 }
-  ]
+    {
+      id: "1",
+      semesterPrice: 1299.0,
+      yearPrice: 2499.0,
+      mealsPerWeek: 12,
+      description: "12 meals per week for a balanced plan.",
+    },
+    {
+      id: "2",
+      semesterPrice: 1599.0,
+      yearPrice: 3099.0,
+      mealsPerWeek: 24,
+      description: "24 meals per week for higher meal consumption.",
+    },
+    {
+      id: "3",
+      semesterPrice: 1399.0,
+      yearPrice: 2699.0,
+      mealsPerWeek: 12,
+      description: "Flexible plan with 12 meals per week.",
+    },
+    {
+      id: "4",
+      semesterPrice: 1499.0,
+      yearPrice: 2899.0,
+      mealsPerWeek: 12,
+      description: "Premium plan with additional benefits.",
+    },
+  ];
 
   React.useEffect(() => {
-    const initialTypes: Record<string, 'Year' | 'Semester'> = {}
-    mealPlans.forEach(plan => {
-      initialTypes[plan.id] = 'Year'
-    })
-    setSelectedTypes(initialTypes)
-  }, [])
+    const initialTypes: Record<string, "Year" | "Semester"> = {};
+    mealPlans.forEach((plan) => {
+      initialTypes[plan.id] = "Year";
+    });
+    setSelectedTypes(initialTypes);
+  }, []);
 
-  const handleTypeChange = (planId: string, type: 'Year' | 'Semester') => {
-    setSelectedTypes(prev => ({
+  const handleTypeChange = (planId: string, type: "Year" | "Semester") => {
+    setSelectedTypes((prev) => ({
       ...prev,
-      [planId]: type
-    }))
-  }
+      [planId]: type,
+    }));
+  };
 
   const getPrice = (plan: MealPlanOption) => {
-    return selectedTypes[plan.id] === 'Year' ? plan.yearPrice : plan.semesterPrice
-  }
+    return selectedTypes[plan.id] === "Year"
+      ? plan.yearPrice
+      : plan.semesterPrice;
+  };
 
   const handleCheckout = (plan: MealPlanOption) => {
-    const price = getPrice(plan)
-    router.push(`/checkout?price=${price}&type=meal-plan&planId=${plan.id}&planType=${selectedTypes[plan.id] || 'Year'}&from=dining/dining-page`)
-  }
+    const price = getPrice(plan);
+    router.push(
+      `/checkout?price=${price}&type=meal-plan&planId=${plan.id}&planType=${
+        selectedTypes[plan.id] || "Year"
+      }&from=dining/dining-page`
+    );
+  };
 
   return (
-      <div className="flex flex-col min-h-screen bg-gray-50">
-        <Header title="Meal Plans" isHomeScreen={false} />
+    <div className="flex flex-col min-h-screen bg-gray-50">
+      <Header title="Meal Plans" isHomeScreen={false} />
 
-        <main className="flex-1 p-4 space-y-6 pt-16">
-          <Card className="bg-white shadow-sm border">
-            <CardContent className="p-4">
-              <h2 className="text-[#8B1A1A] font-semibold text-lg mb-2">Your Current Meal Plan</h2>
-              <p className="text-gray-600">Some information about your meal plan.</p>
-              <p className="text-gray-600">blah blah.</p>
-            </CardContent>
-          </Card>
+      <main className="flex-1 p-4 space-y-6 pt-20 pb-24">
+        <Card className="bg-white shadow-sm border rounded-xl">
+          <CardContent className="p-4">
+            <h2 className="text-[#8B1A1A] font-semibold text-xl mb-3">
+              Your Current Meal Plan
+            </h2>
+            <div className="space-y-2">
+              <p className="text-gray-700">
+                <span className="font-medium">Plan:</span> 12 Meals Per Week
+              </p>
+              <p className="text-gray-700">
+                <span className="font-medium">Remaining Balance:</span> $750.25
+              </p>
+              <p className="text-gray-700">
+                <span className="font-medium">Meals Left This Week:</span> 8
+              </p>
+              <p className="text-gray-700">
+                <span className="font-medium">Plan Expires:</span> May 15, 2025
+              </p>
+            </div>
+            <p className="mt-4 text-sm text-gray-600">
+              Visit the dining hall or use the mobile app to check your daily
+              meal credits and special offers.
+            </p>
+          </CardContent>
+        </Card>
 
-          <div className="grid grid-cols-2 gap-4">
-            {mealPlans.map((plan) => (
-                <Card key={plan.id} className="bg-[#8B1A1A] text-white">
-                  <CardContent className="p-4 flex flex-col h-full">
-                    <div className="mb-4">
-                      <h3 className="text-lg font-semibold">Meals</h3>
-                      <p className="text-2xl font-bold">${getPrice(plan).toFixed(2)}</p>
+        <div className="space-y-4">
+          {mealPlans.map((plan) => (
+            <Card
+              key={plan.id}
+              className="bg-[#841414] text-white shadow-lg hover:shadow-xl transition-shadow duration-300 rounded-3xl overflow-hidden"
+            >
+              <CardContent className="p-0">
+                <div className="p-6 pb-0">
+                  <div className="flex justify-between items-start mb-6">
+                    <div>
+                      <h3 className="text-2xl font-bold mb-1">Meals</h3>
+                      <p className="text-4xl font-bold">
+                        ${getPrice(plan).toFixed(2)}
+                      </p>
                     </div>
-
-                    <div className="flex items-center mb-4">
-                      <CreditCard className="w-5 h-5 mr-2" />
+                    <div className="flex items-center text-lg">
+                      <CreditCard className="w-6 h-6 mr-3" />
                       <span>{plan.mealsPerWeek} Meals Per Week</span>
                     </div>
+                  </div>
+                  <div className="flex items-start mb-4">
+                    <Info className="w-5 h-5 mr-2 flex-shrink-0 mt-1" />
+                    <p className="text-sm">{plan.description}</p>
+                  </div>
+                </div>
 
-                    <div className="flex mb-4 bg-[#6B1414] rounded-lg p-1">
-                      <Button
-                          variant="ghost"
-                          className={`flex-1 ${selectedTypes[plan.id] === 'Year' ? 'bg-[#98D8AA] text-black' : 'text-white'}`}
-                          onClick={() => handleTypeChange(plan.id, 'Year')}
-                      >
-                        Year
-                      </Button>
-                      <Button
-                          variant="ghost"
-                          className={`flex-1 ${selectedTypes[plan.id] === 'Semester' ? 'bg-[#98D8AA] text-black' : 'text-white'}`}
-                          onClick={() => handleTypeChange(plan.id, 'Semester')}
-                      >
-                        Semester
-                      </Button>
-                    </div>
+                <div className="flex border-t border-b border-white">
+                  <button
+                    onClick={() => handleTypeChange(plan.id, "Year")}
+                    className={`flex-1 py-4 px-6 text-xl font-medium flex items-center justify-center gap-2 transition-all duration-300 ease-in-out text-white
+                      ${
+                        selectedTypes[plan.id] === "Year"
+                          ? "bg-[#BFA87C]"
+                          : "bg-[#841414] hover:bg-[#9a1818]"
+                      }`}
+                  >
+                    Year
+                    {selectedTypes[plan.id] === "Year" && (
+                      <Check className="h-5 w-5 ml-2" />
+                    )}
+                  </button>
+                  <button
+                    onClick={() => handleTypeChange(plan.id, "Semester")}
+                    className={`flex-1 py-4 px-6 text-xl font-medium flex items-center justify-center gap-2 transition-all duration-300 ease-in-out border-l border-white text-white
+                      ${
+                        selectedTypes[plan.id] === "Semester"
+                          ? "bg-[#BFA87C]"
+                          : "bg-[#841414] hover:bg-[#9a1818]"
+                      }`}
+                  >
+                    Semester
+                    {selectedTypes[plan.id] === "Semester" && (
+                      <Check className="h-5 w-5 ml-2" />
+                    )}
+                  </button>
+                </div>
 
-                    <Button
-                        className="mt-auto bg-[#98D8AA] hover:bg-[#98D8AA]/90 text-black"
-                        onClick={() => handleCheckout(plan)}
-                    >
-                      Checkout
-                    </Button>
-                  </CardContent>
-                </Card>
-            ))}
-          </div>
-        </main>
+                <button
+                  className="w-full bg-[#841414] hover:bg-[#9a1818] text-white text-xl py-4 px-6 border-t border-white transition-all duration-300 ease-in-out flex items-center justify-center rounded-b-3xl"
+                  onClick={() => handleCheckout(plan)}
+                >
+                  Checkout
+                </button>
+              </CardContent>
+            </Card>
+          ))}
+        </div>
+      </main>
 
-        <BottomNavBar />
-      </div>
-  )
+      <BottomNavBar />
+    </div>
+  );
 }
