@@ -1,4 +1,4 @@
-import React from "react";
+import React, { useEffect, useState } from "react";
 import { Input } from "@/components/ui/input";
 import Header from "@/components/ui/pageHeader";
 import NavBar from "@/components/ui/navBar";
@@ -7,8 +7,10 @@ import IdCard from "@/components/ui/idCard";
 import { Button } from "@/components/ui/button";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import { MapPin, Phone, Mail } from "lucide-react";
+import { currentUserId } from "@/lib/currentUser";
 
 interface ProfileData {
+  id: number;
   name: string;
   idNumber: string;
   email: string;
@@ -18,14 +20,57 @@ interface ProfileData {
 }
 
 export default function ProfilePage() {
-  const profileData: ProfileData = {
-    name: "John Doe",
-    idNumber: "#123456789",
-    email: "johndoe@cofc.edu",
-    address: "123 Sesame Street",
-    phoneNumber: "(678) 999-8212",
-    photoUrl: "/Student_ID_Photo.jpg",
+  const [profileData, setProfileData] = useState<ProfileData | null>(null);
+
+  useEffect(() => {
+    const fetchProfileData = async () => {
+      try {
+        const response = await fetch(`/api/users?userId=${currentUserId}`);
+        if (response.ok) {
+          const data = await response.json();
+          setProfileData(data);
+        } else {
+          console.error('Failed to fetch profile data');
+        }
+      } catch (error) {
+        console.error('Error fetching profile data:', error);
+      }
+    };
+
+    fetchProfileData();
+  }, []);
+
+  const handleFormSubmit = async (e: React.FormEvent) => {
+    e.preventDefault();
+    if (!profileData) return;
+
+    try {
+      const response = await fetch(`/api/users`, {
+        method: 'PUT',
+        headers: {
+          'Content-Type': 'application/json',
+        },
+        body: JSON.stringify(profileData),
+      });
+
+      if (response.ok) {
+        console.log('Profile updated successfully');
+      } else {
+        console.error('Failed to update profile');
+      }
+    } catch (error) {
+      console.error('Error updating profile:', error);
+    }
   };
+
+  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+    const { id, value } = e.target;
+    setProfileData((prevData) => prevData ? { ...prevData, [id]: value } : null);
+  };
+
+  if (!profileData) {
+    return <div>Loading...</div>;
+  }
 
   return (
     <div className="flex flex-col min-h-screen bg-gray-50">
@@ -68,13 +113,14 @@ export default function ProfilePage() {
             <CardTitle>Edit Profile</CardTitle>
           </CardHeader>
           <CardContent>
-            <form className="space-y-4">
+            <form className="space-y-4" onSubmit={handleFormSubmit}>
               <div className="space-y-2">
                 <Label htmlFor="name">Full Name</Label>
                 <Input
                   id="name"
                   defaultValue={profileData.name}
                   className="bg-white"
+                  onChange={handleInputChange}
                 />
               </div>
               <div className="space-y-2">
@@ -84,6 +130,7 @@ export default function ProfilePage() {
                   type="email"
                   defaultValue={profileData.email}
                   className="bg-white"
+                  onChange={handleInputChange}
                 />
               </div>
               <div className="space-y-2">
@@ -92,6 +139,7 @@ export default function ProfilePage() {
                   id="address"
                   defaultValue={profileData.address}
                   className="bg-white"
+                  onChange={handleInputChange}
                 />
               </div>
               <div className="space-y-2">
@@ -100,9 +148,10 @@ export default function ProfilePage() {
                   id="phoneNumber"
                   defaultValue={profileData.phoneNumber}
                   className="bg-white"
+                  onChange={handleInputChange}
                 />
               </div>
-              <Button className="w-full bg-[#841414] hover:bg-[#9a1818]">
+              <Button type="submit" className="w-full bg-[#841414] hover:bg-[#9a1818]">
                 Save Changes
               </Button>
             </form>
